requirements.txt
aniso8601==7.0.0
attrs==19.1.0
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
pyrsistent==0.15.2
python-dotenv==0.10.3
pytz==2019.1
six==1.12.0
Werkzeug==0.15.4
